# webroot/swagger/data/index.yaml
openapi: 3.1.0
info:
  title: 'CB1 Offers API'
  contact:
    email: gho-mansur@yandex.ru
  version: '0.1'
servers:
  -
    url: 'http://cb1.local'
paths:
  /api/offers:
    get:
      tags:
        - offers
      summary: 'Просмотр всех офферов'
      description: 'Просмотр всех офферов'
      operationId: get-offers-all
      responses:
        '200':
          description: 'Просмотр всех офферов'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Offer'
        '404':
          description: 'Нет данных'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
    post:
      tags:
        - offers
      summary: 'Добавление оффера'
      description: 'Добавление оффера'
      operationId: add-offer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddOffer'
      responses:
        '201':
          description: Принято
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  '/api/offers/{guid}':
    get:
      tags:
        - offers
      summary: 'Просмотр одного оффера'
      description: 'Просмотр оффера по guid'
      operationId: get-offer
      responses:
        '200':
          description: 'Просмотр оффера'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offer'
        '404':
          description: 'Оффер на найден'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
    put:
      tags:
        - offers
      summary: 'Редактирование оффера'
      description: 'Добавление оффера'
      operationId: edit-offer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddOffer'
      responses:
        '201':
          description: Принято
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
    delete:
      tags:
        - offers
      summary: 'Удаление оффера'
      description: 'Удалить оффер по guid'
      operationId: delete-offer
      responses:
        '200':
          description: 'Оффер удалён'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: 'Оффер на найден'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
components:
  schemas:
    AddOffer:
      description: 'Description of Offer'
      properties:
        type:
          description: 'Тип сделки'
          type: string
          enum:
            - продажа
            - аренда
        propertyType:
          description: 'Тип недвижимости'
          type: string
          enum:
            - жилая
        category:
          description: 'Категория объекта'
          type: string
          enum:
            - дача
            - дом
            - 'дом с участком'
            - участок
            - 'часть дома'
            - квартира
            - комната
            - таунхаус
            - участок
            - дуплекс
            - гараж
        garageType:
          description: 'Категория гаража'
          type:
            - string
            - 'null'
          enum:
            - гараж
            - машиноместо
            - бокс
        lotNumber:
          description: 'Номер лота'
          type:
            - string
            - 'null'
        cadastralNumber:
          description: 'Кадастровый номер объекта недвижимости'
          type:
            - string
            - 'null'
        url:
          description: 'URL страницы с объявлением на вашем сайте'
          type:
            - string
            - 'null'
        location:
          $ref: '#/components/schemas/Location'
          description: 'Расположение объекта'
      type: object
    Location:
      description: 'Description of Location'
      properties:
        country:
          description: Страна
          type: string
        region:
          description: 'Название субъекта РФ'
          type:
            - string
            - 'null'
        district:
          description: 'Название района субъекта РФ'
          type: string
        localityName:
          description: 'Название населенного пункта'
          type: string
        subLocalityName:
          description: 'Район в населенном пункте'
          type:
            - string
            - 'null'
        address:
          description: 'Адрес объекта — улица и номер здания'
          type: string
        apartment:
          description: 'Номер квартиры'
          type: string
      type: object
    Offer:
      description: 'Description of Offer'
      type: object
      allOf:
        -
          $ref: '#/components/schemas/AddOffer'
        -
          properties:
            internalId:
              description: 'Дата создания объявления'
              type: string
            type:
              description: 'Тип сделки'
              type: string
              enum:
                - продажа
                - аренда
            propertyType:
              description: 'Тип недвижимости'
              type: string
              enum:
                - жилая
            category:
              description: 'Категория объекта'
              type: string
              enum:
                - дача
                - дом
                - 'дом с участком'
                - участок
                - 'часть дома'
                - квартира
                - комната
                - таунхаус
                - участок
                - дуплекс
                - гараж
            garageType:
              description: 'Категория гаража'
              type:
                - string
                - 'null'
              enum:
                - гараж
                - машиноместо
                - бокс
            lotNumber:
              description: 'Номер лота'
              type:
                - string
                - 'null'
            cadastralNumber:
              description: 'Кадастровый номер объекта недвижимости'
              type:
                - string
                - 'null'
            url:
              description: 'URL страницы с объявлением на вашем сайте'
              type:
                - string
                - 'null'
            creationDate:
              description: 'Дата создания объявления'
              type: string
            location:
              $ref: '#/components/schemas/Location'
              description: 'Расположение объекта'
          type: object
    MessageResponse:
      description: 'Description of MessageResponse'
      properties:
        message:
          description: Сообщение
          type: string
      type: object
tags:
  -
    name: offers
    description: offers